Move range indicator logic into new CHUDRecon class
