+++ src/xrGame/HUDTarget.cpp
@@ -29,34 +29,7 @@
 #include <ai/monsters/poltergeist/poltergeist.h>
 
 
-u32 C_ON_ENEMY D3DCOLOR_RGBA(0xff, 0, 0, 0x80);
-u32 C_ON_NEUTRAL D3DCOLOR_RGBA(0xff, 0xff, 0x80, 0x80);
-u32 C_ON_FRIEND D3DCOLOR_RGBA(0, 0xff, 0, 0x80);
-
 #define C_DEFAULT	D3DCOLOR_RGBA(0xff,0xff,0xff,0x80)
-
-#define SHOW_INFO_SPEED		0.5f
-#define HIDE_INFO_SPEED		10.f
-
-static float recon_mindist()
-{
-	return 2.f;
-}
-
-static float recon_maxdist()
-{
-	return 50.f;
-}
-
-static float recon_minspeed()
-{
-	return 0.5f;
-}
-
-static float recon_maxspeed()
-{
-	return 10.f;
-}
 
 static float lerp(float a, float b, float t)
 {
@@ -66,8 +39,6 @@ static float lerp(float a, float b, float t)
 
 CHUDTarget::CHUDTarget()
 {
-	fuzzyShowInfo = 0.f;
-
 	crosshairPos = Fvector();
 	crosshairOpacity = 1.f;
 
@@ -206,141 +177,15 @@ void CHUDTarget::Render()
 	if (HUD().FireposActive())
 		mat_aim.mulB_43(Fmatrix().setHPB(0, 0, hpb_barrel.z - hpb_cam.z));
 
-	float result_dist = GetUIDist();
-
-	Fvector4 pt = Fvector4();
-	if (HUD().FireposActive())
-	{
-		Device.mFullTransform.transform(pt, mat_aim.c);
-		pt.y = -pt.y;
-	}
-
 	// Crosshair color
 	u32 color_readout = C_DEFAULT;
 
-	// Readout font
-	CGameFont* F = UI().Font().pFontGraffiti19Russian;
-	F->SetAligment(CGameFont::alCenter);
-	F->OutSetI(pt.x, pt.y + 0.05f);
-
-	if (psHUD_Flags.test(HUD_CROSSHAIR_DIST))
-		F->OutSkip();
-
-	if (psHUD_Flags.test(HUD_INFO))
-	{
-		bool const is_poltergeist = pp.result.O && !!smart_cast<CPoltergeist*>(pp.result.O);
-
-		if ((pp.result.O && pp.result.O->getVisible()) || is_poltergeist)
-		{
-			CEntityAlive* EA = smart_cast<CEntityAlive*>(pp.result.O);
-			CEntityAlive* pCurEnt = smart_cast<CEntityAlive*>(Level().CurrentEntity());
-			PIItem l_pI = smart_cast<PIItem>(pp.result.O);
-
-			if (IsGameTypeSingle())
-			{
-				CInventoryOwner* our_inv_owner = smart_cast<CInventoryOwner*>(pCurEnt);
-
-				if (EA && EA->g_Alive() && EA->cast_base_monster())
-				{
-					color_readout = C_ON_ENEMY;
-				}
-				else if (EA && EA->g_Alive() && !EA->cast_base_monster())
-				{
-					CInventoryOwner* others_inv_owner = smart_cast<CInventoryOwner*>(EA);
-
-					if (our_inv_owner && others_inv_owner)
-					{
-						switch (RELATION_REGISTRY().GetRelationType(others_inv_owner, our_inv_owner))
-						{
-						case ALife::eRelationTypeEnemy:
-							color_readout = C_ON_ENEMY;
-							break;
-						case ALife::eRelationTypeNeutral:
-							color_readout = C_ON_NEUTRAL;
-							break;
-						case ALife::eRelationTypeFriend:
-							color_readout = C_ON_FRIEND;
-							break;
-						}
-
-						if (fuzzyShowInfo > 0.5f)
-						{
-							CStringTable strtbl;
-							F->SetColor(subst_alpha(color_readout, u8(iFloor(255.f * (fuzzyShowInfo - 0.5f) * 2.f))));
-							F->OutNext("%s", *strtbl.translate(others_inv_owner->Name()));
-							F->OutNext("%s", *strtbl.translate(others_inv_owner->CharacterInfo().Community().id()));
-						}
-					}
-
-					fuzzyShowInfo += SHOW_INFO_SPEED * Device.fTimeDelta;
-				}
-				else if (l_pI && our_inv_owner && result_dist < 2.0f * 2.0f)
-				{
-					if (fuzzyShowInfo > 0.5f && l_pI->NameItem())
-					{
-						F->SetColor(subst_alpha(color_readout, u8(iFloor(255.f * (fuzzyShowInfo - 0.5f) * 2.f))));
-						F->OutNext("%s", l_pI->NameItem());
-					}
-					fuzzyShowInfo += SHOW_INFO_SPEED * Device.fTimeDelta;
-				}
-			}
-			else
-			{
-				if (EA && (EA->GetfHealth() > 0))
-				{
-					if (pCurEnt && GameID() & eGameIDSingle)
-					{
-						if (GameID() & eGameIDDeathmatch) color_readout = C_ON_ENEMY;
-						else
-						{
-							if (EA->g_Team() != pCurEnt->g_Team()) color_readout = C_ON_ENEMY;
-							else color_readout = C_ON_FRIEND;
-						};
-						if (result_dist >= recon_mindist() && result_dist <= recon_maxdist())
-						{
-							float ddist = (result_dist - recon_mindist()) / (recon_maxdist() - recon_mindist());
-							float dspeed = recon_minspeed() + (recon_maxspeed() - recon_minspeed()) * ddist;
-							fuzzyShowInfo += Device.fTimeDelta / dspeed;
-						}
-						else
-						{
-							if (result_dist < recon_mindist())
-								fuzzyShowInfo += recon_minspeed() * Device.fTimeDelta;
-							else
-								fuzzyShowInfo = 0;
-						};
-
-						if (fuzzyShowInfo > 0.5f)
-						{
-							clamp(fuzzyShowInfo, 0.f, 1.f);
-							int alpha_C = iFloor(255.f * (fuzzyShowInfo - 0.5f) * 2.f);
-							u8 alpha_b = u8(alpha_C & 0x00ff);
-							F->SetColor(subst_alpha(color_readout, alpha_b));
-							F->OutNext("%s", *pp.result.O->cName());
-						}
-					}
-				};
-			};
-		}
-		else
-		{
-			fuzzyShowInfo -= HIDE_INFO_SPEED * Device.fTimeDelta;
-		}
-		clamp(fuzzyShowInfo, 0.f, 1.f);
-	}
-
-	if (psHUD_Flags.test(HUD_CROSSHAIR_DIST))
-	{
-		F->OutSetI(pt.x, pt.y + 0.05f);
-		F->SetColor(color_readout);
-#ifdef DEBUG
-		F->OutNext("%4.1f - %4.2f - %d", result_dist, PP.power, PP.pass);
-#else
-		F->OutNext("%4.1f", result_dist);
-#endif
-	}
+	HUDRecon.SetTransform(mat_aim);
+	HUDRecon.SetColor(color_readout);
+	HUDRecon.OnRender(GetUIDist());
 
 	// Use the crosshair color unless the readout color is non-default
+	color_readout = HUDRecon.GetColor();
 	u32 color_crosshair = color_readout == C_DEFAULT ? g_crosshair_color : color_readout;
 
 	// Modulate color alpha
